guzzle split
